Split daemon error types into seperate module
The main module should only be daemon code, not also errors.
